% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.fit.R
\name{mvp.fit}
\alias{mvp.fit}
\title{mvp.fit}
\usage{
mvp.fit(X, Z = NULL, N1 = 100, N2 = 100)
}
\arguments{
\item{X}{n-by-m matrix of n m-variate Poisson samples}

\item{Z}{n-by-m matrix of covariates for the mean latent rates}

\item{N1}{number of samples}

\item{N2}{number of steps in Metropolis Hastings}
}
\description{
mvp.fit
}

