% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.get.Y1.R
\name{mvp.get.Y1}
\alias{mvp.get.Y1}
\title{mvp.get.Y1}
\usage{
mvp.get.Y1(X, Y2, A)
}
\arguments{
\item{X}{n-by-m matrix of n m-variate Poisson samples}

\item{Y2}{n-by-m(m-1)/2 matrix of latent Poisson covariance samples}

\item{A}{m-by-m(m+1)/2 matrix from X = AY}
}
\value{
n-by-m matrix of latent Poisson mean samples
}
\description{
mvp.get.Y1
}

