% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.matrix.R
\name{mvp.matrix}
\alias{mvp.matrix}
\title{mvp.matrix}
\usage{
mvp.matrix(m)
}
\arguments{
\item{m}{dimension of the m-variate Poisson}
}
\value{
the MVP matrix A from X = AY (Karlis 2005), where X is MVP-distributed, and Y is a vector of independent Poissons.
}
\description{
mvp.matrix
}

