% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.prob.R
\name{mvp.prob}
\alias{mvp.prob}
\title{mvp.prob}
\usage{
mvp.prob(x, theta, method, logarithm, n.mc = 10000)
}
\arguments{
\item{x}{MVP sample}

\item{theta}{underlying Poisson rates}

\item{n.mc}{number of Monte Carlo samples if method='MC"}
}
\value{
Estimate of the probability P(x) for a given vector theta of Poisson rates
}
\description{
mvp.prob
}

