% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.sample.R
\name{mvp.sample}
\alias{mvp.sample}
\title{mvp.sample}
\usage{
mvp.sample(n, theta.means, theta.covs)
}
\arguments{
\item{n}{number of samples to be drawn}

\item{theta.means}{underlying Poisson rates for the mean}

\item{theta.covs}{underlying Poisson rates for the covariance}
}
\value{
n-by-m matrix of n m-variate Poisson samples, where m is of size y1
}
\description{
mvp.sample
}

