% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.fit.R
\name{mvp.fit}
\alias{mvp.fit}
\title{mvp.fit}
\usage{
mvp.fit(method, X, Z = NULL, verbose = FALSE, ...)
}
\arguments{
\item{method}{"HMC" or "Gibbs"}

\item{X}{n-by-m matrix of n m-variate Poisson samples}

\item{Z}{n-by-m matrix of covariates for the mean latent rates}
}
\value{
results of the sampling method
}
\description{
mvp.fit
}

